backend/worker/requirements.txt
celery==5.4.0
redis==5.2.1

SQLAlchemy==2.0.38
psycopg[binary]==3.2.5
pgvector==0.3.6

boto3==1.34.162
pypdf==5.2.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.6
pydantic-settings==2.7.1

